\name{annot_HGU133A}
\docType{data}
\alias{annot_HGU133A}
\title{annot\_HGU133A}
\description{Partial Affymetrix annotation for HG-U133A. Used in the example cross-species analysis presented in the vignette of package annotationTools.}
\format{data.frame with 188 rows and 43 columns}
\source{\url{http://www.affymetrix.com}}
\keyword{datasets}
