\name{gene_orthologs_part}
\docType{data}
\alias{gene_orthologs_part}
\title{Example gene_orthologs file}
\description{Shortened gene_orthologs database file.
}
\format{A short version of 'gene_orthologs' table containing 2 ortholog clusters. Ortholog clusters in 'gene_orthologs' derive from NCBI's Eukaryotic Genome Annotation pipeline (see https://www.ncbi.nlm.nih.gov/kis/info/how-are-orthologs-calculated/).}
\source{\url{ftp.ncbi.nlm.nih.gov/gene/DATA/gene_orthologs.gz}}
\keyword{datasets}
