\name{HG-U133_Plus_2_annot_part}
\docType{data}
\alias{HG-U133_Plus_2_annot_part}
\title{Example Affymetrix annotation file}
\description{Truncated Affymetrix annotation file for GeneChip array HG-U133 Plus 2.
}
\format{A table containing 7 probe set IDs and the associated annotation.}
\source{\url{http://www.affymetrix.com}}
\keyword{datasets}
