\name{HG-U133_Plus_2_ortholog_part}
\docType{data}
\alias{HG-U133_Plus_2_ortholog_part}
\title{Example Affymetrix ortholog file}
\description{Truncated Affymetrix ortholog file for GeneChip array HG-U133 Plus 2.
}
\format{A table containing 5 probe set IDs and their orthologs probe set IDs on other GeneChip arrays.}
\source{\url{http://www.affymetrix.com}}
\keyword{datasets}
