\name{Mouse430_2_annot_part}
\docType{data}
\alias{Mouse430_2_annot_part}
\title{Example Affymetrix annotation file}
\description{Truncated Affymetrix annotation file for GeneChip array Mouse 430 2.
}
\format{A table containing 9 probe set IDs and the associated annotation.}
\source{\url{http://www.affymetrix.com}}
\keyword{datasets}
