\name{ortho}
\docType{data}
\alias{ortho}
\title{ortho}
\description{Partial ortholog probe sets mapping table of MG-U74Av2 probe sets to HG-U133A probe sets. Generated with ps2ps.R and Affymetrix annotations for MG-U74Av2 and HG-U133A arrays. Used in the example cross-species analysis presented in the vignette of package annotationTools.}
\format{data.frame with 100 rows and 4 columns.}
\source{\url{http://www.affymetrix.com} for Affymetrix annotations}
\keyword{datasets}
