\name{table_human}
\docType{data}
\alias{table_human}
\title{table\_human}
\description{Partial list of probe sets from differential expression analysis of Huntington's disease patients versus controls. Used in the example cross-species analysis presented in the vignette of package annotationTools.}
\format{data.frame with 188 rows (probe sets) and 12 columns.}
\source{\url{http://hdbase.org/cgi-bin/welcome.cgi}}
\keyword{datasets}
