\name{table_mouse}
\docType{data}
\alias{table_mouse}
\title{table\_mouse}
\description{Top 100 probe sets from the differential expression analysis of CHL2 mice versus wild-type mice (MG-U74Av2). Used in the example cross-species analysis presented in the vignette of package annotationTools.}
\format{data.frame with 100 rows and 8 columns}
\source{\url{http://hdbase.org/cgi-bin/welcome.cgi}}
\keyword{datasets}
