\name{gene_group_part}
\docType{data}
\alias{gene_group_part}
\title{Example gene_group file}
\description{Shortened gene_group database file.
}
\format{A short version of 'gene_group' table containing 2 ortholog clusters. 'gene_group' is the flat file version of NCBI's 'Orthologs from Annotation pipeline' database (see https://www.ncbi.nlm.nih.gov/news/03-13-2014-gene-provides-orthologs-regions/).}
\source{\url{ftp.ncbi.nlm.nih.gov/gene/DATA/gene_group.gz}}
\keyword{datasets}
